// web/src/api/model/getParents200Item.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * school-manager-api
 * OpenAPI spec version: 1.0
 */
import type { GetParents200ItemStudentsItem } from './getParents200ItemStudentsItem';
import type { GetParents200ItemType } from './getParents200ItemType';

export type GetParents200Item = {
    email: string;
    id: string;
    name: string;
    pamId?: string;
    password: string;
    phone: string;
    students: GetParents200ItemStudentsItem[];
    surname: string;
    type: GetParents200ItemType;
};